\name{filter_indel}
\alias{filter_indel}
\title{
Filter somatic indels.
}
\description{
Apply the standard set of indel filters on a dataframe generated by the ShlienLab SSM pipeline.
}
\usage{
filter_indel(data = NULL, source = "WXS", coverage = TRUE, indels = c("ins", "del"))
}
\arguments{
  \item{data}{R dataframe from the ShlienLab SSM pipeline (required).}
  \item{source}{Type of library sequenced (default: WXS -- Must be one of WGS, WXS, CPANEL).}
  \item{coverage}{Filter based on coverage criteria from library type defined in source (default: TRUE).}
  \item{indels}{Vector containing labels for indels (default: c('ins', 'del')).}
}
\value{
}
\author{
Richard J. de Borja
}
\keyword{ ssm }% use one of  RShowDoc("KEYWORDS")
\keyword{ indel }
\keyword{ filter }
