\name{filter_snv}
\alias{filter_snv}
\title{
Filter somatic single nucleotide variants.
}
\description{
Apply the standard set of single nucleotide variant filters on a dataframe generated by the ShlienLab SSM pipeline.
}
\usage{
filter_snv(data = NULL, coverage = TRUE, source = "WXS", exac = 0.001, vaf = 0)
}
\arguments{
  \item{data}{R dataframe from the ShlienLab SSM pipeline (required).}
  \item{coverage}{Filter based on coverage criteria from library type defined in source (default: TRUE).}
  \item{source}{Type of library sequenced (default: WXS -- Must be one of WGS, WXS, CPANEL)}
  \item{exac}{ExAC threshold (default: 0.01)}
  \item{vaf}{VAF threshold (default: 0.0)}
}
\value{
Returns a dataframe containing only SNV mutations that have been filtered using the standard set of lab filters.
}
\author{
Richard J. de Borja
}
\keyword{ ssm }
\keyword{ snv }
\keyword{ filter }
